fix: load external app only if needed
